refactor: split parsers/factory.ts into config, providers, boardingPass, email modules
